refactor: mediainfo module
