+++ animepipeline/mediainfo/__init__.py
@@ -1,1 +1,3 @@
-from animepipeline.mediainfo.mediainfomini import gen_file_name, FileNameInfo, rename_file, get_media_info  # noqa
+from animepipeline.mediainfo.mediainfo import get_media_info  # noqa
+from animepipeline.mediainfo.name import gen_file_name, rename_file  # noqa
+from animepipeline.mediainfo.type import FileNameInfo, MediaInfo  # noqa
